feat: Footer layout 적용
