+++ ticatch-front/app/(with-header)/layout.tsx
@@ -1,10 +1,12 @@
 import Header from '@components/Header';
+import Footer from '@components/Footer';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div className="container">
       <Header />
       {children}
+      <Footer />
     </div>
   );
 }
